Updated all projects to Swift 1.2; added example second level to project 8.

+++ project14-files/Helper.swift
@@ -11,7 +11,7 @@ import UIKit
 
 func RandomInt(#min: Int, #max: Int) -> Int {
 	if max < min { return min }
-	return Int(arc4random_uniform((max - min) + 1)) + min
+	return Int(arc4random_uniform(UInt32((max - min) + 1))) + min
 }
 
 func RandomFloat() -> Float {
